pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitcloakd"
version = "1.0.3"
description = "Cloak Git repositories with GPG encryption - hide your code in plain sight"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "HaKC.dev", email = "dev@hakc.dev"}
]
maintainers = [
    {name = "HaKC.dev", email = "dev@hakc.dev"}
]
keywords = [
    "git", "encryption", "gpg", "pgp", "security", "github",
    "git-crypt", "secrets", "repository", "cli", "cloak", "privacy"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Version Control :: Git",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "python-gnupg>=0.5.0",
    "GitPython>=3.1.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "watchdog>=3.0.0",
    "colorama>=0.4.6",
    "hakcer>=1.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
gitcloakd = "gitcloakd.cli.main:cli"
gitcloak = "gitcloakd.cli.main:cli"

[project.urls]
Homepage = "https://github.com/haKC-ai/gitcloakd"
Documentation = "https://github.com/haKC-ai/gitcloakd#readme"
Repository = "https://github.com/haKC-ai/gitcloakd"
Issues = "https://github.com/haKC-ai/gitcloakd/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gitcloakd = ["templates/*", "data/*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=gitcloakd"
